Revert "refactor: fix filename and folder structure"
This reverts commit 84ed6d525023c7edbc9561c9b6ac05472ef3d7e2.
